requirements.txt
gdown>=4.7.1
pytesseract>=0.3.10
Spire.Pdf>=8.7.2
opencv-python-headless>=4.8.0
pandas>=2.0.3
python-docx>=0.8.11
pyyaml>=6.0.1
numpy>=1.24.3
python-dateutil>=2.8.2

